% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_persons.R
\name{get_persons}
\alias{get_persons}
\title{Get person objects from BibEntry or bibentry object}
\usage{
get_persons(bib, data = NULL, exclude_crossref = FALSE)
}
\arguments{
\item{bib}{A BibEntry or bibentry object that may be subselected
from a larger set}

\item{exclude_crossref}{The InProceedings entries for this dataset reference
their respective Proceedings entries. RefManageR::ReadBib combines the parent
fields when the dataset is built. Setting exclude_crossref = TRUE prevents
the parent person fields from being duplicated for the InProceedings entries.}

\item{refbib}{A BibEntry or bibentry object that contains any
crossreferences relevant to the function, such as exclude_crossref. Required
for excluding persons inherited from crossref entries.}
}
\value{
A data.frame with columns: family, given, role, key, and crossref
}
\description{
A function to return persons from BibEntry and bibentry objects
}
\examples{
obj <- icusbib[author = "Silva", year = "1988"]
obj
get_persons(obj , data = icusbib, exclude_crossref = FALSE)
get_persons(obj , data = icusbib, exclude_crossref = TRUE)
obj <- icusbib[list(author = "Kubler"), list(author="^Oates")]
obj
get_persons(obj , data = icusbib, exclude_crossref = FALSE)
get_persons(obj , data = icusbib, exclude_crossref = TRUE)
}
\seealso{
\link{icusbib}
}
