% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{package}
\name{icus.data}
\alias{icus.data}
\alias{icus.data-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
icus.data
A data package for icus.data.

Use \code{data(package='icus.data')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{bibset}
\link{nobels}
\link{participants}
\link{programs}
\link{role_defs}
\link{whoswho}
}
