% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{nobels}
\alias{nobels}
\title{Nobel Laureates who participated in ICUS}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{ID}{ID for participants, relational to \code{participants} table}
\item{Surname}{}
\item{Given}{}
\item{Year}{Year in which participant receive Nobel Prize}
\item{Category}{Category of prize}
\item{OrgCountry}{Country of recipient's organization}
}
}
\source{
{The data comes from Nobel Media AB.}
}
\description{
In his 1977 criticism of ICUS IV,
Irving Horowitz noted the prominence of Nobel
Laureates on the program and a striking fall-off
in talent beyond a stellar list of
sponsors. This table is based on the Nobel Media AB
dataset.
}
\details{
Detailed description of the data
}
\examples{
library(icus.data)
nobels
}
\seealso{
\link{bibset}
\link{icus.data}
\link{participants}
\link{programs}
\link{role_defs}
\link{whoswho}
}
