% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{participants}
\alias{participants}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{ID}{}
\item{Surname}{}
\item{Given}{}
\item{Conference}{}
\item{Role}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{icus.data}
\link{bibset}
\link{nobels}
\link{programs}
\link{role_defs}
\link{whoswho}
}
