% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{programs}
\alias{programs}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{Conference}{}
\item{Year}{}
\item{Site}{}
\item{City}{}
\item{Country}{}
\item{ProgTitle}{}
\item{Count}{}
\item{Nations}{}
\item{ProgPages}{}
\item{Cover}{}
\item{Title}{}
\item{Foreword}{}
\item{Purpose}{}
\item{Executive}{}
\item{TOC}{}
\item{Structure}{}
\item{Schedule}{}
\item{Participants}{}
\item{Sponsor}{}
\item{Obituary}{}
\item{Venue}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{icus.data}
\link{bibset}
\link{nobels}
\link{participants}
\link{role_defs}
\link{whoswho}
}
