% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{role_defs}
\alias{role_defs}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{Abbr}{}
\item{Role}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{icus.data}
\link{bibset}
\link{nobels}
\link{participants}
\link{programs}
\link{whoswho}
}
