% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{viaf}
\alias{viaf}
\title{VIAF Identifiers for ICUS participants}
\format{
A \code{tbl_df} containing the following fields:
\describe{
\item{ID}{ICUS ID assigned for this data package.}
\item{Surname}{Curated surname.}
\item{Given}{Curated given name.}
\item{VIAF}{List of VIAF persistent identifiers for person.}
}
}
\source{
Manually compiled from VIAF with addition searches to
disambiguate names where necessary.
}
\description{
Persistent identifiers manually curated from
the Virtual International Authorities File at viaf.org.
}
\details{
Detailed description of the data
}
\examples{
library(icus.data)
library(viafr)
}
\seealso{
\link{icus.data}
\link{icusbib}
\link{nobels}
\link{participants}
\link{programs}
\link{role_defs}
\link{whoswho}
}
