% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{whoswho}
\alias{whoswho}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{Surname}{}
\item{Given}{}
\item{variable}{}
\item{value}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{icus.data}
}
