% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{package}
\name{icus.data}
\alias{icus.data}
\alias{icus.data-package}
\title{Dataset for International Conferences on the Unity of Science}
\description{
More than 2000 scientists and scholars participated in the 22
International Conferences on the Unity of the Sciences (ICUS)
held between 1972 and 2000. The main spreadsheet in this dataset
is a table of 2045 people who have been
listed in ICUS programs as speakers, moderators, discussants,
advisors and other committee members. This dataset contains
tables related to researching these conferences.
}
\details{
icus.data
A data package for icus.data.

Use \code{data(package='icus.data')$results[, 3]} to see
a list of available data sets in this data package.
}
\seealso{
\link{bibset}
\link{nobels}
\link{participants}
\link{programs}
\link{role_defs}
\link{whoswho}
}
