% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{participants}
\alias{participants}
\title{ICUS participants and their roles}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{ID}{Format: ICUSxxxx. Assigned sequentially to rows in
\code{ICUSRoles.xlsx}.
This is the authoratitive relational column. Other tables are
copied from here.}
\item{Surname}{}
\item{Given}{}
\item{Conference}{Format: ICUSxx }
\item{Role}{Single character codes. Definitions in role_defs.}
}
}
\source{
The data comes from ICUS programs.
}
\description{
Pariticpants in ICUS conferences were manually compiled from
ICUS programs. See the vignette
\code{icus.data::icus-program-observations}
for observations about the 22 conference programs. During compilation,
the various roles of the participants were entered for each conference.
The resulting Excel file is incuded in this package in folder \code{extdata}.
}
\details{
Detailed description of the data

The original data is in wide format with ICUSxx as column
headers and
the cells filled with character strings consisting of the roles the named
participant filled in the conference. Only one letter per type of role even
though a participant may have filled a role more that once in a
conference.
This is most common for discussant (X) roles. The original table can
be viewed by running the examples. The \code{data.table}
\code{participants} is in long
format with the role strings also pivoted.
}
\examples{
## Load and view original roles from \code{ICUSRoles.xlsx}
xlsx <- file.path(system.file(package="icus.data"),
                  "extdata","ICUSRoles.xlsx")
library(readxl)
library(data.table)
excel_sheets(xlsx)
roles <- as.data.table(read_xlsx(xlsx, sheet="CrossChart"))
roles
##
library(icus.data)
participants
}
\seealso{
\link{bibset}
\link{icus.data}
\link{nobels}
\link{programs}
\link{role_defs}
\link{whoswho}
}
