% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{viaf}
\alias{viaf}
\title{VIAF Identifiers for ICUS participants}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{ID}{}
\item{Surname}{}
\item{Given}{}
\item{VIAF}{}
}
}
\source{
Manually compiled.
}
\description{
Persistent identifiers manually curated from
the Virtual International Authorities File at viaf.org.
}
\details{
Detailed description of the data
}
\seealso{
\link{icus.data}
\link{icusbib}
\link{nobels}
\link{participants}
\link{programs}
\link{role_defs}
\link{whoswho}
}
