% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icus.data.R
\docType{data}
\name{whoswho}
\alias{whoswho}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{ID}{}
\item{Surname}{}
\item{Given}{}
\item{Edition}{}
\item{Birthdate}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{icus.data}
\link{bibset}
\link{nobels}
\link{participants}
\link{programs}
\link{role_defs}
}
